+++ Changelog.md
@@ -1,5 +1,6 @@
 # 0.5.0 - 2015-12-8
 
+- Support `x-summary` and `x-description` in Path Item Objects
 - Generate Swagger schema and spec validation annotations.
 - Implement support for link relations in parser annotations.
 - Encode dashes (`-`) in URI template parameters.